% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_svm}
\alias{read_svm}
\title{Read SVM Light sparse data}
\usage{
read_svm(path, zero_indexing = TRUE, zero_one_response = FALSE,
  transpose = TRUE, nfeatures = 0L)
}
\arguments{
\item{path}{: Path of file saved in .svm format}

\item{zero_indexing}{: integer indicating if stored data indicated columns indexing starting from 0 or 1}

\item{zero_one_response}{: teh default output for the response variable is 1, -1; set to true to change to 1, 0.}

\item{transpose}{: logical indicating if feature matrix should be returned transposed (see details).}

\item{nfeatures}{: optional non negative integer indicating the total number of features in the matrix, if the value is zero (by default)
the number of features is taken from the observed data. Use when the number of features is known and you want to run data in batches.}
}
\value{
an object of class dgCMatrix from the Matrix package
}
\description{
Convenience function to read svm light data format
}
\details{
Most sparse matrix libraries have better support for column-compressed sparse data format while algorithms 
require to read information by row. The default return_transpose = true to make this easier.
}
