\name{readsvm-package}
\alias{readsvm-package}
\alias{readsvm}
\docType{package}
\title{
\packageTitle{readsvm}
}
\description{
\packageDescription{readsvm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{readsvm}
\packageIndices{readsvm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{readsvm}

Maintainer: \packageMaintainer{readsvm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
